+++ Tweak.h
@@ -203,15 +203,15 @@
     WKWebView *_webView;
 }
 - (void)moveByOffset:(NSInteger)offset;
-- (id)positionFromPosition:(id)arg1 inDirection:(long long)arg2 offset:(long long)arg3;
-- (id)positionFromPosition:(id)arg1 toBoundary:(long long)arg2 inDirection:(long long)arg3;
+- (id)positionFromPosition:(id)arg1 inDirection:(NSInteger)arg2 offset:(NSInteger)arg3;
+- (id)positionFromPosition:(id)arg1 toBoundary:(NSInteger)arg2 inDirection:(NSInteger)arg3;
 - (id)textRangeFromPosition:(id)arg1 toPosition:(id)arg2;
 - (void)setSelectedTextRange:(UITextRange *)arg1;
 - (id)selectedText;
 - (void)selectWordBackward;
 - (UITextRange *)selectedTextRange;
 - (void)setSelectedTextRange:(UITextRange *)arg1;
-- (long long)baseWritingDirectionForPosition:(id)arg1 inDirection:(long long)arg2;
+- (NSInteger)baseWritingDirectionForPosition:(id)arg1 inDirection:(NSInteger)arg2;
 - (id<UITextInputTokenizer>)tokenizer;
 - (UITextRange *)markedTextRange;
 - (id)textInRange:(id)arg1;
@@ -220,13 +220,13 @@
 - (NSRange)_markedTextNSRange;
 - (UITextPosition *)beginningOfDocument;
 - (id)selectionRectsForRange:(id)arg1;
-- (long long)comparePosition:(id)arg1 toPosition:(id)arg2;
-- (void)beginSelectionInDirection:(long long)arg1 completionHandler:(id)arg2;
-- (void)selectTextWithGranularity:(long long)arg1 atPoint:(CGPoint)arg2 completionHandler:(id)arg3;
+- (NSInteger)comparePosition:(id)arg1 toPosition:(id)arg2;
+- (void)beginSelectionInDirection:(NSInteger)arg1 completionHandler:(id)arg2;
+- (void)selectTextWithGranularity:(NSInteger)arg1 atPoint:(CGPoint)arg2 completionHandler:(id)arg3;
 - (void)updateSelectionWithExtentPoint:(CGPoint)arg1 completionHandler:(/*^block */id)arg2;
-- (void)updateSelectionWithExtentPoint:(CGPoint)arg1 withBoundary:(long long)arg2 completionHandler:(id)arg3;
+- (void)updateSelectionWithExtentPoint:(CGPoint)arg1 withBoundary:(NSInteger)arg2 completionHandler:(id)arg3;
 - (id)webView;
-- (id)positionFromPosition:(id)arg1 offset:(long long)arg2;
+- (id)positionFromPosition:(id)arg1 offset:(NSInteger)arg2;
 
 - (id)_moveToEndOfWord:(BOOL)arg1 withHistory:(id)arg2;
 - (id)_moveToEndOfLine:(BOOL)arg1 withHistory:(id)arg2;
@@ -270,9 +270,9 @@
 @property (assign,nonatomic) BOOL isNone;                                        
 @property (assign,nonatomic) BOOL isRange;                                       
 @property (assign,nonatomic) BOOL isEditable;                                    
-@property (assign,nonatomic) long long selectedTextLength;
+@property (assign,nonatomic) NSInteger selectedTextLength;
 @property (nonatomic,copy) NSArray *selectionRects;
-+ (id)textRangeWithState:(BOOL)arg1 isRange:(BOOL)arg2 isEditable:(BOOL)arg3 startRect:(CGRect)arg4 endRect:(CGRect)arg5 selectionRects:(id)arg6 selectedTextLength:(unsigned long long)arg7;
++ (id)textRangeWithState:(BOOL)arg1 isRange:(BOOL)arg2 isEditable:(BOOL)arg3 startRect:(CGRect)arg4 endRect:(CGRect)arg5 selectionRects:(id)arg6 selectedTextLength:(NSUInteger)arg7;
 - (BOOL)isEqual:(id)arg1;
 - (id)description;
 - (BOOL)isEmpty;
@@ -289,11 +289,11 @@
 - (void)setIsEditable:(BOOL)arg1;
 - (void)setIsNone:(BOOL)arg1;
 - (void)setIsRange:(BOOL)arg1;
-- (void)setSelectedTextLength:(long long)arg1;
+- (void)setSelectedTextLength:(NSInteger)arg1;
 - (void)setSelectionRects:(NSArray *)arg1;
 - (BOOL)isRange;
 - (BOOL)isNone;
-- (long long)selectedTextLength;
+- (NSInteger)selectedTextLength;
 @end
